Robot is implemented as a cell now
Robot now is nested in class board. The reason is following:
Every time we move the robot needs to have the reference of the board
and the robot state.

but i dont think we should build a robot class.

and the reason why i use cell type is it needs a position. 

The last thing is the class doubly linked list, it is not finished. For
now, just put it there.