board should be fine for now
to initialize the board we should use randomSetDirt(int) and
randomSetObstacle(int). But for testing reasons, we can add dirt by
calling function setDirt(int, int).

We may wanna add obstacles, but I cant see the reason now. So the
function is now there. If we do need it later. Then I will put it there.